refactor: Reorganize project structure and consolidate under app directory
- Move components, utils, and lib directories under app/
- Update import paths to reflect new directory structure
- Remove unused smtp-message component
- Update tsconfig and next.config paths to align with new structure
- Clean up error handling in middleware
