+++ app/components/hero.tsx
